feature: get document path
This commit implements the api endpoint in web in order to fetch the
document. It includes abstractions to easily develop api endpoints and
removes the old api project because we don't need it now that we are not
a latency-sensitive proxy anymore.
