--- apps/api/README.md
@@ -1,29 +1,0 @@
-## Available Scripts
-
-### `pnpm run dev`
-
-Run the server in development mode.
-
-### `pnpm test`
-
-Run all unit-tests with hot-reloading.
-
-### `pnpm run test:watch`
-
-Run all unit-tests without hot-reloading.
-
-### `pnpm run lint`
-
-Check for linting errors.
-
-### `pnpm run build`
-
-Build the project for production.
-
-### `pnpm start`
-
-Run the production build (Must be built first).
-
-## Additional Notes
-
-- If `pnpm run dev` gives you issues with bcrypt on MacOS you may need to run: `pnpm rebuild bcrypt --build-from-source`.
