--- apps/api/src/common/Paths.ts
@@ -1,17 +1,0 @@
-/**
- * Express router paths go here.
- */
-
-export default {
-  Base: '',
-  Api: {
-    Base: '/api',
-    V1: {
-      Base: '/v1',
-      Chat: {
-        Base: '/chat',
-        Completions: '/completions',
-      },
-    },
-  },
-} as const
