--- apps/api/src/common/misc.ts
@@ -1,5 +1,0 @@
-export enum NodeEnvs {
-  Dev = 'development',
-  Test = 'test',
-  Production = 'production',
-}
