--- apps/api/src/index.ts
@@ -1,7 +1,0 @@
-import EnvVars from '@src/common/EnvVars'
-
-import server from './server'
-
-const SERVER_START_MSG = 'Express server started on port: ' + EnvVars.PORT
-
-server.listen(EnvVars.PORT, () => console.info(SERVER_START_MSG))
