--- apps/api/src/routes/api/v1/chat/index.ts
@@ -1,10 +1,0 @@
-import Paths from '@src/common/Paths'
-import { Router } from 'express'
-
-import { completions } from './routes'
-
-const router = Router()
-
-router.use(Paths.Api.V1.Chat.Completions, completions)
-
-export default router
