--- apps/api/src/routes/api/v1/chat/routes.ts
@@ -1,6 +1,0 @@
-import HttpStatusCodes from '@src/common/HttpStatusCodes'
-import { Request, Response } from 'express'
-
-export function completions(_: Request, res: Response) {
-  return res.status(HttpStatusCodes.OK).json()
-}
