--- apps/api/src/routes/api/v1/index.ts
@@ -1,10 +1,0 @@
-import Paths from '@src/common/Paths'
-import { Router } from 'express'
-
-import chatRouter from './chat'
-
-const router = Router()
-
-router.use(Paths.Api.V1.Chat.Base, chatRouter)
-
-export default router
