--- apps/api/src/services/elastic/client.ts
@@ -1,10 +1,0 @@
-import { Client } from '@elastic/elasticsearch'
-import EnvVars from '@src/common/EnvVars'
-
-export default new Client({
-  node: EnvVars.ELASTIC_URL,
-  auth: {
-    username: EnvVars.ELASTIC_USERNAME,
-    password: EnvVars.ELASTIC_PASSWORD,
-  },
-})
