--- apps/api/test/support/Paths.ts
@@ -1,8 +1,0 @@
-/**
- * Convert paths to full for testing.
- */
-
-import Paths from '@src/common/Paths'
-import jetPaths from 'jet-paths'
-
-export default jetPaths(Paths)
