--- apps/api/test/support/jasmine.json
@@ -1,7 +1,0 @@
-{
-  "spec_dir": "test",
-  "spec_files": ["**/*test.ts"],
-  "helpers": ["helpers/**/*.js"],
-  "stopSpecOnExpectationFailure": false,
-  "random": true
-}
