--- apps/api/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@latitude-data/typescript-config/base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "baseUrl": "./",
-    "paths": {
-      "@src/*": ["src/*"],
-      "$core/*": ["../../packages/core/src/*"]
-    },
-    "useUnknownInCatchVariables": false
-  },
-  "include": ["src/**/*.ts", "test/**/*.ts"]
-}
