+++ packages/core/src/data-access/documentVersions/index.ts
@@ -1,0 +1,3 @@
+export * from './getDocumentsAtCommit'
+export * from './getDocumentByPath'
+export * from './getDocumentById'
