+++ packages/core/src/data-access/index.ts
@@ -1,8 +1,5 @@
-import { Database } from '@latitude-data/core'
-
 export * from './users'
 export * from './projects'
 export * from './commits'
 export * from './documentVersions'
-
-export type AppContext = { db: Database }
+export * from './apiKeys'
