+++ packages/core/src/services/commits/index.ts
@@ -1,0 +1,2 @@
+export * from './create'
+export * from './merge'
