+++ packages/core/src/services/index.ts
@@ -1,3 +1,5 @@
 export * from './users'
 export * from './workspaces'
 export * from './documentVersions'
+export * from './commits'
+export * from './projects'
